deleted CustomerForm.kt and added its funcionality into CustomerFragment.kt
